add an email validator utility class

+++ src/main/java/edu/icet/ems/repository/EmployeeJpaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface EmployeeJpaRepository extends CrudRepository<EmployeeEntity, Integer> {
+public interface EmployeeJpaRepository extends CrudRepository<EmployeeEntity, Long> {
 
     Optional<EmployeeEntity> findByEmail(String email);
 }
